+++ media_kit_libs_windows_audio/windows/media_kit_libs_windows_audio_plugin_c_api.cpp
@@ -1,6 +1,7 @@
 #include "include/media_kit_libs_windows_audio/media_kit_libs_windows_audio_plugin_c_api.h"
 
 #include <flutter/plugin_registrar_windows.h>
+#include <iostream>
 
 void MediaKitLibsWindowsAudioPluginCApiRegisterWithRegistrar(
     FlutterDesktopPluginRegistrarRef registrar) {