+++ media_kit_test/lib/tests/06.paint_first_frame.dart
@@ -3,9 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:media_kit/media_kit.dart';
 import 'package:media_kit_video/media_kit_video.dart';
 
+import '../common/globals.dart';
 import '../common/sources/sources.dart';
-
-// ignore_for_file: use_build_context_synchronously
 
 Future<void> paintFirstFrame(BuildContext context) async {
   // Create [Player] and [VideoController] instances.
@@ -17,33 +16,53 @@ Future<void> paintFirstFrame(BuildContext context) async {
     Player(),
   ];
   List<VideoController> controllers = [
-    await VideoController.create(players[0]),
-    await VideoController.create(players[1]),
-    await VideoController.create(players[2]),
-    await VideoController.create(players[3]),
-    await VideoController.create(players[4]),
+    await VideoController.create(
+      players[0],
+      enableHardwareAcceleration: enableHardwareAcceleration.value,
+    ),
+    await VideoController.create(
+      players[1],
+      enableHardwareAcceleration: enableHardwareAcceleration.value,
+    ),
+    await VideoController.create(
+      players[2],
+      enableHardwareAcceleration: enableHardwareAcceleration.value,
+    ),
+    await VideoController.create(
+      players[3],
+      enableHardwareAcceleration: enableHardwareAcceleration.value,
+    ),
+    await VideoController.create(
+      players[4],
+      enableHardwareAcceleration: enableHardwareAcceleration.value,
+    ),
   ];
   // Open some [Playable]s.
   // Do not start playback i.e. play: false.
   for (int i = 0; i < 5; i++) {
     await players[i].open(
-      Media(sources[i % sources.length]),
+      Playlist(
+        sources.map((e) => Media(e)).toList(),
+        index: i,
+      ),
       play: false,
     );
   }
 
-  // Some voluntary delay.
-  await Future.delayed(const Duration(seconds: 1));
+  await Future.wait(controllers.map((e) => e.waitUntilFirstFrameRendered));
 
   // The first frame should be painted!
-  await Navigator.of(context).push(
-    MaterialPageRoute(
-      builder: (context) => PaintFirstFrameScreen(
-        players: players,
-        controllers: controllers,
+  if (context.mounted) {
+    await Navigator.of(context).push(
+      MaterialPageRoute(
+        builder: (context) => PaintFirstFrameScreen(
+          players: players,
+          controllers: controllers,
+        ),
       ),
-    ),
-  );
+    );
+  }
+
   for (final e in controllers) {
     await e.dispose();
   }