+++ media_kit_video/windows/video_output.h
@@ -54,6 +54,7 @@ class VideoOutput {
   VideoOutput(int64_t handle,
               std::optional<int64_t> width,
               std::optional<int64_t> height,
+              bool enable_hardware_acceleration,
               flutter::PluginRegistrarWindows* registrar,
               ThreadPool* thread_pool_ref);
 
@@ -81,10 +82,16 @@ class VideoOutput {
   mpv_render_context* render_context_ = nullptr;
   std::optional<int64_t> height_ = std::nullopt;
   std::optional<int64_t> width_ = std::nullopt;
+  bool enable_hardware_acceleration_ = true;
   int64_t texture_id_ = 0;
   flutter::PluginRegistrarWindows* registrar_ = nullptr;
   ThreadPool* thread_pool_ref_ = nullptr;
+  // For preventing any asynchronous operations (primarily texture objects
+  // deletion after unregister in |Resize|) access this object after
+  // destruction.
   bool destroyed_ = false;
+
+  std::mutex textures_mutex_ = std::mutex();
 
   std::unordered_map<int64_t, std::unique_ptr<flutter::TextureVariant>>
       texture_variants_ = {};