+++ media_kit_libs_android_audio/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.3
+
+- build: bump dependencies
+
+## 1.0.2
+
+- perf: enable `extractNativeLibs`
+- fix: package:ffmpeg_kit_flutter compatibility
+
 ## 1.0.0
 
-- Initial release.
+- Initial release