+++ media_kit_libs_macos_video/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.5
+
+- build: bump min macOS version to 10.9
+
+## 1.0.4
+
+- build: bump libs version
+
 ## 1.0.3
 
 - fix: update xcframeworks